requirements.txt
absl-py==2.3.1
attrs==25.4.0
cffi==2.0.0
contourpy==1.3.3
cycler==0.12.1
flatbuffers==25.12.19
fonttools==4.61.1
jax==0.9.0
jaxlib==0.9.0
joblib==1.5.3
kiwisolver==1.4.9
matplotlib==3.10.8
mediapipe==0.10.13
ml_dtypes==0.5.4
numpy==2.4.1
opencv-contrib-python==4.13.0.90
opencv-python==4.13.0.90
opt_einsum==3.4.0
packaging==26.0
Panda3D==1.10.16
panda3d-gltf==1.3.0
panda3d-simplepbr==0.13.1
pillow==12.1.0
protobuf==4.25.8
pycparser==3.0
pygame==2.6.1
pyparsing==3.3.2
pyperclip==1.11.0
python-dateutil==2.9.0.post0
scikit-learn==1.7.2
scipy==1.17.0
screeninfo==0.8.1
six==1.17.0
sounddevice==0.5.5
sswg==4.3.0
threadpoolctl==3.6.0
typing_extensions==4.15.0
ursina==8.3.0
